moved to a scripted tile system

+++ Assets/Scripts/GameScripts/GameEventDispatcher.cs
@@ -7,7 +7,7 @@ public static class GameEventDispatcher
     public delegate void PlayerOnLavaEvent();
     public static event PlayerOnLavaEvent PlayerOnLava;
 
-    public delegate void PlayerReachedStarEvent(int x, int y);
+    public delegate void PlayerReachedStarEvent(float x, float y);
     public static event PlayerReachedStarEvent PlayerReachedStar;
 
     public static void DispatchPlayerAtExitEvent()
@@ -20,7 +20,7 @@ public static class GameEventDispatcher
         PlayerOnLava?.Invoke();
     }
 
-    public static void DispatchPlayerReachedStarEvent(int x, int y)
+    public static void DispatchPlayerReachedStarEvent(float x, float y)
     {
         PlayerReachedStar?.Invoke(x, y);
     }
